Fix SkyBlockProfile to store all skill XP as double (not long) — update the field type, getter/setter, and serialisation/deserialisation so (#2310)
By Jarvis agent Vega.

+++ plugin/src/main/java/com/skyblock/plugin/listener/MiningListener.java
@@ -48,9 +48,9 @@ public final class MiningListener implements Listener {
         Player player = event.getPlayer();
         SkyBlockProfile profile = ProfileManager.getInstance()
                 .getOrCreateProfile(player.getUniqueId());
-        int before = SKILL_MANAGER.levelForXp("mining", profile.getSkillXp("mining"));
+        int before = SKILL_MANAGER.levelForXp("mining", (long) profile.getSkillXp("mining"));
         profile.addSkillXp("mining", xp);
-        int after = SKILL_MANAGER.levelForXp("mining", profile.getSkillXp("mining"));
+        int after = SKILL_MANAGER.levelForXp("mining", (long) profile.getSkillXp("mining"));
         if (after > before) {
             player.sendTitle("§aSkill Level Up!", "§eMining §a→ §eLVL " + after, 10, 60, 20);
         }
